sécurisation de la suppression

+++ controller/astreController.js
@@ -62,7 +62,23 @@ async function loadFixtures() {
     return await connexion.saveObject(astres, connexion.astresUrl)
 }
 
-async function deleteAstre(astreID) {
+async function deleteAstre(astreID, token) {
+    if ( !astreID){
+        console.log("erreur controleur : astre object need an _id field to be updated")
+        return {code: "400", message: "astre object need an _id field to be updated"}
+    }
+
+    let astreToDelete = await this.getAstreByID(astreID)
+    if ( !astreToDelete || astreToDelete === {} ){
+        console.log("erreur controleur : given id doesn't match to any existing astre")
+        return {code: "404", message: "given id doesn't match to any existing astre"}
+    }
+
+    let loginUser = await utilisateurController.extractLoginFromToken(token)
+    if ( ! astreToDelete.auteur || astreToDelete.auteur !== loginUser){
+        console.log("erreur controleur : you need to be the author of an astre to update it")
+        return {code: "404", message: "you need to be the author of an astre to update it"}
+    }
     return await connexion.deleteObject(astreID, connexion.astresUrl)
 }
